feat: Reestructure UI to Field Management services. (#866)
* feat: Reestructure UI to Field Management services.

* change search end-points.
